Minería de datos/Semana 08/data-dogs/requirements.txt
requests==2.32.3
pandas==2.3.3
python-dotenv==1.2.2
matplotlib==3.10.8
openpyxl==3.1.5
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
alembic==1.18.4
streamlit==1.55.0
plotly==6.6.0
numpy==2.4.3
packaging==23.2
wheel==0.45.1